Move series volume middlewares to their own file
